+++ app/controllers/catch_all_controller.rb
@@ -6,7 +6,7 @@ class CatchAllController < ActionController::Base
       redirect.used!
       redirect_to redirect.destination_uri, status: redirect.status_code
     else
-      render text: 'Page not found', status: 404
+      render plain: 'Page not found', status: 404
     end
   end
 end